Removes redundant sourceSets block (Gradle defaults), collapses JUnit 5 deps to the aggregator artifact, drops explicit kotlin-stdlib (added automatically by the Kotlin plugin), and bump Kotlin from 1.9.20 to 2.3.21.

+++ src/main/java/com/vahoss/Test.kt
@@ -149,7 +149,7 @@ object Reader {
                 if (size == -1) return NC
                 bId = 0
             }
-            return buf[bId++].toChar()
+            return buf[bId++].toInt().toChar()
         }
 
     fun nextInt(): Int {
